put settings items in a composable
